+++ tedxoxford/src/components/Accordion.js
@@ -6,10 +6,10 @@ const Accordion = ({title, content}) => {
     return (
         <div className="accordion-item">
             <div className="accordion-title" onClick={() => setIsActive(!isActive)}>
-                <div>{title}</div>
+                <div dangerouslySetInnerHTML={ { __html: title } }></div>
                 <div>{isActive ? '-' : '+'}</div>
             </div>
-            {isActive && <div className="accordion-content">{content}</div>}
+            {isActive && <div className="accordion-content" dangerouslySetInnerHTML={ { __html: content } }></div>}
         </div>
     )
 }